feat: ResponseEntity added on NinjaController and MissionsController

+++ src/main/java/br/com/NinjaRegistration/NinjaRegistration/Missions/MissionsController.java
@@ -2,6 +2,8 @@ package br.com.NinjaRegistration.NinjaRegistration.Missions;
 
 import br.com.NinjaRegistration.NinjaRegistration.Ninjas.NinjaDTO;
 import br.com.NinjaRegistration.NinjaRegistration.Ninjas.NinjaModel;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,8 +18,9 @@ public class MissionsController {
 
     //create Missions (CREATE)
     @PostMapping("/create")
-    public MissionsDTO createMissions(@RequestBody MissionsDTO missionsDTO){
-        return missionsService.create(missionsDTO);
+    public ResponseEntity<String> createMissions(@RequestBody MissionsDTO missionsDTO){
+        missionsService.create(missionsDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Mission '" + missionsDTO.getName() + "' created");
     }
 
     @GetMapping("/print/all")
@@ -37,13 +40,25 @@ public class MissionsController {
 
     //Modify mission by id (UPDATE)
     @PutMapping("/update/{id}")
-    public MissionsDTO update(@RequestBody MissionsDTO missionsDTO, @PathVariable Long id){
-        return missionsService.update(missionsDTO, id);
+    public ResponseEntity<String> update(@RequestBody MissionsDTO missionsDTO, @PathVariable Long id){
+        if(printById(id) != null){
+            missionsService.update(missionsDTO, id);
+            return ResponseEntity.status(HttpStatus.OK).body("Mission " + printById(id).getName() + " data updated");
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Mission " + printById(id).getName() + " was not found and also not updated");
+        }
     }
 
     //Delete missions by id (DELETE)
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id){
-        missionsService.delete(id);
+    public ResponseEntity<String> delete(@PathVariable Long id){
+        if (missionsService.printById(id) != null) {
+            missionsService.delete(id);
+            return ResponseEntity.ok().body("ID: " + id +", Mission " + printById(id).getName() + " deleted");
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Mission id: " + id + " not found");
+        }
     }
 }
